refactor(dashboard): extract the panel stack section
Second composition-section extraction (B1): the prompt/skill/todo panel
stack moves into src/dashboard/sections/panelStack.ts. The prompt-store
acquisition await stays at the composition root so the extraction is
microtask-identical (the startup diagnostics contract pins the exact event
order); the todo-command and skill-wiring content anchors follow the
constructions into the section file. No behavior change.
